+++ src/main/java/com/wenxin2/marioverse/datagen/ItemTagsGen.java
@@ -130,6 +130,7 @@ public class ItemTagsGen extends ItemTagsProvider {
 
         tag(Tags.Items.BUCKETS_ENTITY_WATER)
                 .add(ItemRegistry.CHEEP_CHEEP_BUCKET.get())
+                .add(ItemRegistry.EEP_CHEEP_BUCKET.get())
                 .add(ItemRegistry.SPINY_CHEEP_CHEEP_BUCKET.get());
 
         tag(Tags.Items.BUCKETS_POWDER_SNOW)
@@ -255,6 +256,7 @@ public class ItemTagsGen extends ItemTagsProvider {
         tag(TagRegistry.CHEEP_CHEEP_ITEMS)
                 .add(ItemRegistry.CHEEP_CHEEP.get())
                 .add(ItemRegistry.COLD_CHEEP_CHEEP.get())
+                .add(ItemRegistry.EEP_CHEEP.get())
                 .add(ItemRegistry.WARM_CHEEP_CHEEP.get());
 
         tag(TagRegistry.CRAFTS_WARP_DOOR)