+++ src/main/java/com/wenxin2/marioverse/event_handlers/RegistryEventHandlers.java
@@ -281,6 +281,7 @@ public class RegistryEventHandlers {
             newEntries.add(LootItem.lootTableItem(ItemRegistry.CHEEP_CHEEP.get()).setWeight(20).build());
             newEntries.add(LootItem.lootTableItem(ItemRegistry.COLD_CHEEP_CHEEP.get()).setWeight(15).build());
             newEntries.add(LootItem.lootTableItem(ItemRegistry.WARM_CHEEP_CHEEP.get()).setWeight(15).build());
+            newEntries.add(LootItem.lootTableItem(ItemRegistry.EEP_CHEEP.get()).setWeight(15).build());
             newEntries.add(LootItem.lootTableItem(ItemRegistry.SPINY_CHEEP_CHEEP.get()).setWeight(8).build());
             entriesField.set(pool, newEntries);
         } catch (NoSuchFieldException | IllegalAccessException e) {