+++ changelog.md
@@ -1,14 +1,20 @@
 ## Update 1.10.0
 - Added the Cheep Cheep+
-  - If the player is above water, it will jump out to attack you
+  - Jumps out of water
   - On contact, damages the player
   - Normal Cheep Cheeps spawn anywhere
   - Cold Cheep Cheeps spawn anywhere cold
   - Warm Cheep Cheeps spawn anywhere hot
   - Added Bucket of Cheep Cheep
   - Added Raw & Cooked Cheep Cheep
+- Added the Eep Cheep
+  - Swims away from the player
+  - Jumps out of water
+  - On contact, damages the player
+  - Added Bucket of Eep Cheep
+  - Eep Cheeps spawn anywhere
 - Added the Spiny Cheep Cheep
-  - If the player is above water, it will jump out to attack you
+  - Jumps out of water
   - On contact, damages the player with brief poison
   - Added Bucket of Spiny Cheep Cheep
   - Added Raw & Cooked Spiny Cheep Cheep